backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
google-generativeai==0.7.2
python-dotenv==1.0.1
scikit-learn==1.5.2
joblib==1.4.2
requests==2.32.3
beautifulsoup4==4.12.3
pytesseract==0.3.13
Pillow==10.4.0
nltk==3.9.1
sentence-transformers==3.2.1
